Move RBAC requests to the component
